Fix all clippy warnings across the codebase
- Replace manual range checks with RangeInclusive::contains() in shared
 ELF string literal parser (manual_range_contains)
- Collapse nested if statements in x86 linker weak object handling
 (collapsible_if, 4 instances)
- Remove no-op `<< 0` shift in ARM assembler encoder (identity_op)
- Remove redundant `as u32` casts in ARM ELF writer ADRP/ADR relocation
 handling (unnecessary_cast, 4 instances)
- Remove `(0b00 << 22)` zero-value OR in PMUL encoding (identity_op)

Verified: all tests pass (99.6%+ on all arches, 47/48 projects pass,
mbedtls i686 failure is pre-existing). x86 kernel build passes.

+++ src/backend/arm/assembler/elf_writer.rs
@@ -637,8 +637,8 @@ impl ElfWriter {
                         // ADR: op immlo[1:0] 10000 immhi[18:0] Rd
                         // immlo is bits [30:29], immhi is bits [23:5]
                         let imm = pc_offset as i32;
-                        let immlo = ((imm as u32) & 0x3) as u32;
-                        let immhi = (((imm as u32) >> 2) & 0x7FFFF) as u32;
+                        let immlo = (imm as u32) & 0x3;
+                        let immhi = ((imm as u32) >> 2) & 0x7FFFF;
                         word |= (immlo << 29) | (immhi << 5);
                     }
                     275 => {
@@ -648,8 +648,8 @@ impl ElfWriter {
                         let target_page = (target_offset as i64) & !0xFFF;
                         let page_off = target_page - pc_page;
                         let imm = (page_off >> 12) as i32;
-                        let immlo = ((imm as u32) & 0x3) as u32;
-                        let immhi = (((imm as u32) >> 2) & 0x7FFFF) as u32;
+                        let immlo = (imm as u32) & 0x3;
+                        let immhi = ((imm as u32) >> 2) & 0x7FFFF;
                         word |= (immlo << 29) | (immhi << 5);
                     }
                     _ => {
